requirements.txt
absl-py==2.1.0
accelerate==0.30.0
aiohttp==3.9.3
aiosignal==1.3.1
albumentations==1.4.1
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyio==4.3.0
appdirs==1.4.4
asttokens==2.4.1
attrs==23.2.0
bitsandbytes==0.43.1
blinker==1.7.0
blis==0.7.11
bokeh==3.4.1
braceexpand==0.1.5
Brotli==1.1.0
cachetools==5.3.3
catalogue==2.0.10
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
colorama==0.4.6
comm==0.2.2
confection==0.1.4
contourpy==1.2.1
cryptography==42.0.5
cycler==0.12.1
cymem==2.0.8
dataclasses==0.8
datasets==2.14.4
debugpy==1.8.1
decorator==5.1.1
decord==0.6.0
dill==0.3.7
distro==1.9.0
docker-pycreds==0.4.0
einops==0.7.0
einops-exts==0.0.4
en-core-web-sm==3.7.1
et-xmlfile==1.1.0
exceptiongroup==1.2.0
executing==2.0.1
filelock==3.13.4
fonttools==4.51.0
frozenlist==1.4.1
fsspec==2024.3.1
ftfy==6.1.3
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.29.0
google-auth-oauthlib==1.0.0
grad-cam==1.5.0
grpcio==1.60.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.5
httpx==0.27.0
huggingface_hub==0.22.2
hydra-core==1.3.2
hyperframe==6.0.1
idna==3.6
imagecodecs==2024.1.1
imageio==2.34.0
importlib_metadata==7.1.0
importlib_resources==6.4.0
inflection==0.5.1
iopath==0.1.9
ipykernel==6.29.3
ipython==8.22.2
jedi==0.19.1
Jinja2==3.1.3
joblib==1.4.0
jupyter_client==8.6.1
jupyter_core==5.7.2
kiwisolver==1.4.5
langcodes==3.3.0
lazy_loader==0.4
lightning-bolts==0.7.0
lightning-utilities==0.11.2
llvmlite==0.42.0
Markdown==3.6
MarkupSafe==2.1.5
matplotlib==3.8.4
matplotlib-inline==0.1.6
more-itertools==10.2.0
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.15
munkres==1.1.4
murmurhash==1.0.10
nest_asyncio==1.6.0
networkx==3.3
nibabel==5.2.1
nltk==3.8.1
numba==0.59.1
numpy==1.26.4
oauthlib==3.2.2
omegaconf==2.3.0
open_clip_torch==2.24.0
open-flamingo==0.0.2
openai==1.23.2
opencv-python==4.9.0
openpyxl==3.1.2
packaging==24.0
pandas==2.2.1
parso==0.8.4
pathtools==0.1.2
patsy==0.5.6
peft==0.10.0
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.3.0
pip==24.0
platformdirs==4.2.0
ply==3.11
portalocker==2.8.2
preshed==3.0.9
prompt-toolkit==3.0.42
protobuf==4.25.1
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==15.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.22
pydantic==2.7.0
pydantic_core==2.18.1
pyDeprecate==0.3.2
Pygments==2.17.2
PyJWT==2.8.0
pynndescent==0.5.12
pyOpenSSL==24.0.0
pyparsing==3.1.2
PyQt5==5.15.9
PyQt5-sip==12.12.2
PySocks==1.7.1
python-dateutil==2.9.0
pytorch-lightning==1.7.7
pytz==2024.1
pyu2f==0.1.5
PyWavelets==1.4.1
PyYAML==6.0.1
pyzmq==25.1.2
qudida==0.0.4
regex==2023.12.25
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
safetensors==0.4.2
scikit-image==0.22.0
scikit-learn==1.4.1.post1
scipy==1.13.0
seaborn==0.13.1
sentencepiece==0.1.99
sentry-sdk==1.45.0
setproctitle==1.3.3
setuptools==69.2.0
sip==6.7.12
six==1.16.0
smart-open==6.4.0
smmap==5.0.0
sniffio==1.3.1
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
stack-data==0.6.2
statsmodels==0.14.1
sympy==1.12
tensorboard==2.13.0
tensorboard-data-server==0.7.0
thinc==8.2.3
threadpoolctl==3.4.0
tifffile==2024.2.12
timm==0.9.16
tokenizers==0.15.2
toml==0.10.2
tomli==2.0.1
torch==2.1.2.post303
torchmetrics==0.11.4
torchvision==0.16.1+b88453f
tornado==6.4
tqdm==4.66.2
traitlets==5.14.2
transformers==4.39.3
ttach==0.0.3
typer==0.9.4
typing_extensions==4.11.0
tzdata==2024.1
umap-learn==0.5.5
urllib3==2.2.1
wandb==0.16.5
wasabi==1.1.2
wcwidth==0.2.13
weasel==0.3.4
webdataset==0.2.86
Werkzeug==3.0.2
wheel==0.43.0
xxhash==3.4.1
xyzservices==2024.4.0
yarl==1.9.4
zipp==3.17.0
